Move _register_logfile_on_master to internals.py
